Add abstract class for UI related drawable
